+++ src/main/webapp/consentGroupContainer/ConsentGroupContainer.js
@@ -61,7 +61,7 @@ export const ConsentGroupContainer = hh(class ConsentGroupContainer extends Comp
     if (!$.fn.dataTable.isDataTable("#history-table")) {
       $("#history-table").DataTable({
         dom: '<"H"Tfr><"pull-right"B><div>t</div><"F"lp>',
-        buttons: ['excelHtml5', 'csvHtml5', 'printContent'],
+        buttons: ['excelHtml5', 'csvHtml5', 'printData'],
         language: { search: 'Filter:' },
         pagingType: "full_numbers",
         order: [1, "desc"]
