+++ src/main/webapp/projectContainer/ProjectContainer.js
@@ -66,7 +66,7 @@ export const ProjectContainer = hh(class ProjectContainer extends Component {
     if (!$.fn.dataTable.isDataTable("#history-table")) {
       $("#history-table").DataTable({
         dom: '<"H"Tfr><"pull-right"B><div>t</div><"F"lp>',
-        buttons: ['excelHtml5', 'csvHtml5', 'printContent'],
+        buttons: ['excelHtml5', 'csvHtml5', 'print'],
         language: { search: 'Filter:' },
         pagingType: "full_numbers",
         order: [1, "desc"]
